minor change on columns

+++ src/models/Lembur.js
@@ -14,7 +14,7 @@ Lembur.init(
         allowNull: false,
       },
       status:{
-        type: DataTypes.INTEGER,
+        type: DataTypes.SMALLINT,
         allowNull: false
       },
     },
